Added sprite adding/remove public functions to game engine. Replaced pacman sprite sheet with permanent solution

+++ PacmanRemastered/src/engine/API.java
@@ -30,5 +30,5 @@ public interface API {
      * @param fx Sprite frame x
      * @param fy Sprite frame y
      */
-    public void drawSprite(int index, String picname, int x, int y, int w, int h, int fx, int fy);
+    public void drawSprite(int index, String picname, double x, double y, int w, int h, int fx, int fy);
 }
